Forgotten "long long" instead of uint64_t.
git-svn-id: https://svn.code.sf.net/p/stemkit/code@1019 faf13d3b-af0d-0410-97ba-fcc57a00092e

+++ include/sk/rt/Thread.h
@@ -82,7 +82,7 @@ namespace sk {
         sk::util::Holder<thread::Runner> _runnerHolder;
 
         sk::util::String _name;
-        long long _id;
+        uint64_t _id;
     };
   }
 }
